+++ src/Aws/OpsWorks/Enum/RootDeviceType.php
@@ -14,9 +14,15 @@
  * permissions and limitations under the License.
  */
 
-namespace Aws\OpsWorks\Exception;
+namespace Aws\OpsWorks\Enum;
+
+use Aws\Common\Enum;
 
 /**
- * Indicates that a request was invalid.
+ * Contains enumerable RootDeviceType values
  */
-class ValidationException extends OpsWorksException {}
+class RootDeviceType extends Enum
+{
+    const EBS = 'ebs';
+    const INSTANCE_STORE = 'instance-store';
+}